Move deploy modules to otx
