--- docs/source/guide/reference/mpa/deploy.rst
@@ -1,34 +1,0 @@
-Deploy
-^^^^^^^
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-
-.. automodule:: otx.mpa.deploy
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.apis
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.utils
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.utils.mmdeploy
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.utils.onnx
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.utils.operations_domain
-   :members:
-   :undoc-members:
-
-.. automodule:: otx.mpa.deploy.utils.utils
-   :members:
-   :undoc-members:
