+++ otx/algorithms/common/adapters/mmdeploy/utils/__init__.py
@@ -1,3 +1,4 @@
+"""Init file for otx.algorithms.common.adapters.mmdeploy.utils."""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
