+++ otx/algorithms/common/adapters/mmdeploy/utils/mmdeploy.py
@@ -1,3 +1,4 @@
+"""Functions for mmdeploy adapters."""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -5,14 +6,24 @@
 import importlib
 
 import onnx
-from mmcv.utils import ConfigDict
 
 
 def is_mmdeploy_enabled():
+    """Checks if the 'mmdeploy' Python module is installed and available for use.
+
+    Returns:
+        bool: True if 'mmdeploy' is installed, False otherwise.
+
+    Example:
+        >>> is_mmdeploy_enabled()
+        True
+    """
     return importlib.util.find_spec("mmdeploy") is not None
 
 
 def mmdeploy_init_model_helper(ctx, model_checkpoint=None, cfg_options=None, **kwargs):
+    """Helper function for initializing a model for inference using the 'mmdeploy' library."""
+
     model_builder = kwargs.pop("model_builder")
     model = model_builder(
         ctx.model_cfg,
@@ -31,12 +42,14 @@ def mmdeploy_init_model_helper(ctx, model_checkpoint=None, cfg_options=None, **k
     return model
 
 
-def update_deploy_cfg(onnx_path, deploy_cfg, mo_options={}):
+def update_deploy_cfg(onnx_path, deploy_cfg, mo_options=None):
+    """Update the 'deploy_cfg' configuration file based on the ONNX model specified by 'onnx_path'."""
+
     from mmdeploy.utils import get_backend_config, get_ir_config
 
     onnx_model = onnx.load(onnx_path)
     ir_config = get_ir_config(deploy_cfg)
-    backend_config = get_backend_config(deploy_cfg)
+    _ = get_backend_config(deploy_cfg)
 
     # update input
     input_names = [i.name for i in onnx_model.graph.input]
@@ -47,6 +60,7 @@ def update_deploy_cfg(onnx_path, deploy_cfg, mo_options={}):
     ir_config["output_names"] = output_names
 
     # update mo options
+    mo_options = mo_options if mo_options else dict()
     deploy_cfg.merge_from_dict({"backend_config": {"mo_options": mo_options}})
