+++ otx/algorithms/common/adapters/mmdeploy/utils/operations_domain.py
@@ -1,3 +1,4 @@
+"""Add domina function."""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -6,4 +7,5 @@ DOMAIN_CUSTOM_OPS_NAME = "org.openvinotoolkit"
 
 
 def add_domain(name_operator: str) -> str:
+    """Function for adding to DOMAIN_CUSTOM_OPS_NAME."""
     return DOMAIN_CUSTOM_OPS_NAME + "::" + name_operator
