--- otx/mpa/deploy/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-from .utils import is_mmdeploy_enabled
-
-__all__ = [
-    "is_mmdeploy_enabled",
-]
