+++ otx/mpa/det/exporter.py
@@ -5,7 +5,7 @@
 import numpy as np
 from mmcv.runner import wrap_fp16_model
 
-from otx.mpa.deploy.utils import sync_batchnorm_2_batchnorm
+from otx.algorithms.common.adapters.mmdeploy.utils import sync_batchnorm_2_batchnorm
 from otx.mpa.exporter_mixin import ExporterMixin
 from otx.mpa.registry import STAGES
 from otx.mpa.utils.logger import get_logger
@@ -46,7 +46,7 @@ class DetectionExporter(ExporterMixin, DetectionStage):
         from mmdet.apis.inference import LoadImage
         from mmdet.datasets.pipelines import Compose
 
-        from ..deploy.apis import NaiveExporter
+        from ...algorithms.common.adapters.mmdeploy.apis import NaiveExporter
 
         def get_fake_data(cfg, orig_img_shape=(128, 128, 3)):
             pipeline = [LoadImage()] + cfg.data.test.pipeline[1:]
