+++ tests/unit/algorithms/common/adapters/mmdeploy/utils/test_deploy_utils_operations_domain.py
@@ -2,7 +2,10 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-from otx.mpa.deploy.utils.operations_domain import DOMAIN_CUSTOM_OPS_NAME, add_domain
+from otx.algorithms.common.adapters.mmdeploy.utils.operations_domain import (
+    DOMAIN_CUSTOM_OPS_NAME,
+    add_domain,
+)
 from tests.test_suite.e2e_test_system import e2e_pytest_unit
